requirements.txt
openai>=1.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
sqlalchemy>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pytest>=7.0.0
pypdf>=3.0.0
